Optimized !

+++ lib/src/main/java/com/github/chenyoca/validation/LazyLoader.java
@@ -4,22 +4,22 @@ package com.github.chenyoca.validation;
  * Created by YooJia.Chen
  * YooJia.Chen@gmail.com
  * 2014-07-14
- * Lazy loader of the test extra values
+ * Lazy loader of the test extra value
  */
 public interface LazyLoader {
 
     /**
-     * @return Return the INT values, size of 1 or 2.
+     * @return Return the INT value, size of 1 or 2.
      */
-    int[] intValues();
+    long[] intValues();
 
     /**
-     * @return Return the DOUBLE values, size of 1 or 2.
+     * @return Return the DOUBLE value, size of 1 or 2.
      */
     double[] doubleValues();
 
     /**
-     * @return Return the STRING values, size of 1 or 2.
+     * @return Return the STRING value, size of 1 or 2.
      */
     String[] stringValues();
 }
